Testing all

+++ MGLEngine.Shared/Interfaces/IGraphicLibrary.h
@@ -5,8 +5,8 @@
 #include <vector>
 class IGraphicLibrary {
 	public:
-		virtual void* GetVerticeBuffer(int shaderIndex, size_t sizeInBytes)=0;
-		virtual uint32_t* GetIndicesBuffer(int shaderIndex, size_t sizeInBytes)=0;
+		virtual void* GetVerticeBuffer(size_t shaderIndex, size_t sizeInBytes)=0;
+		virtual uint32_t* GetIndicesBuffer(size_t shaderIndex, size_t sizeInBytes)=0;
 		virtual void FlushVerticeBuffer(int shaderIndex)=0;
 		virtual void FlushIndicesBuffer(int shaderIndex)=0;
 		virtual void Run(std::vector<ShaderContext>& shaders,GlobalBindingsTable& bindingTable) = 0;
